move MainMenu to components due to its more component-like nature
