+++ mobile/components/MainMenu/index.js
@@ -4,16 +4,15 @@ import { compose, graphql, withApollo } from 'react-apollo'
 import { withRouter } from 'react-router'
 import { connect } from 'react-redux'
 import { Animated, Dimensions, Keyboard, View } from 'react-native'
-import Loading from '../../components/Loading'
+import Loading from '../Loading'
 import styles from '../../styles/styles'
 import appStyle from '../../styles/appStyle'
 import levelConfig from 'thebrain-shared/helpers/levelConfig'
-
 import currentUserQuery from 'thebrain-shared/graphql/account/currentUser'
 import sessionCountQuery from 'thebrain-shared/graphql/items/sessionCount'
 import userDetailsQuery from 'thebrain-shared/graphql/userDetails/userDetails'
 
-import WithData from '../../components/WithData'
+import WithData from '../WithData'
 import MainMenuHeader from './components/MainMenuHeader'
 import MainMenuOptions from './components/MainMenuOptions'
